Feature: TextWatcher, CheckBox, Swipe, DB의 기존값 받아오기

+++ app/src/main/java/com/kt_study/todo_alarm/db/CategoryEntity.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "table_category")
 data class CategoryEntity(
     @PrimaryKey
-    val id: Long = 0L,
+    val id: Int = 0,
     var title: String = "",
 )
